+++ uninstaller_safety_other.go
@@ -5,3 +5,7 @@ package main
 func isWindowsUninstallerSafe() bool {
 	return true
 }
+
+func prepareWindowsUninstallerForUpdate() bool {
+	return true
+}